workspace-main/skills/ai-news-fetcher/requirements.txt
lxml>=4.9.0
aiohttp>=3.8.0
pyyaml>=6.0
click>=8.0.0
python-dateutil>=2.8.0
cssselect>=1.2.0
